requirements.txt
lightning
matplotlib
mlflow
numpy
pandas
Pillow
scikit_learn
torch
torchmetrics
torchvision
tqdm
safetensors
seaborn
huggingface_hub